split up and refactor system check page
